Rewrote the interface yet another time...
I'm getting close to what I want it to do. It is now a promise pipeline, the exception handling is not in the steps, but in the Promise.
This makes implementing new steps much easier, since they can now just return a result or throw an exception, and are not called
if the previous step threw an exception